Abstract filtering data from components
