+++ components/Nav/SearchSidebar/SearchSidebar.jsx
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React, { useCallback, useEffect, useMemo, useRef, useState, useReducer } from 'react'
+import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 
 import { default as NavSearchTable } from './SearchTable'
 import PropTypes from 'prop-types'
@@ -24,8 +24,8 @@ import styled from '@emotion/styled'
 import { useRouter } from 'next/router'
 import useUserStore from 'store/use-user-state'
 import { withAlgorandPriceQuery } from '@/hooks'
-import dayjs from 'dayjs'
-import { testnetAssets } from '../../AgeOfProjects'
+import useSearchFilter from '@/hooks/useSearchFilter'
+
 // export const Section = styled.section`
 //   height: inherit;
 //   width: 100%;
@@ -145,56 +145,7 @@ export function NavSearchSidebar({
     [push, handleExternalClick]
   )
 
-  const filterReducer = (state, action) => {
-    switch (action.type) {
-      case 'updateSliderValue':
-        return {
-          ...state,
-          [action.field]: action.value
-        }
-      case 'setPriceMax':
-        return {
-          ...state,
-          priceMax: action.value
-        }
-      case 'toggleMarketCap':
-        return {
-          ...state,
-          isFilteringMarketCap: !state.isFilteringMarketCap
-        }      
-      case 'toggleAgeOfProject':
-        return {
-          ...state,
-          isFilteringAgeOfProject: !state.isFilteringAgeOfProject
-        }      
-      case 'toggleMarketPrice':
-        return {
-          ...state,
-          isFilteringPrice: !state.isFilteringPrice
-        }      
-      case 'toggleNFTOnly':
-        return {
-          ...state,
-          isFilteringNFTOnly: !state.isFilteringNFTOnly
-        }      
-      default:
-        break;
-    }
-    return state
-  }
-  const initialState = {
-    marketCapAmount: 0,
-    isFilteringMarketCap: false,
-    // ageOfProject: 0,
-    ageOfProject: [0, 20],
-    ageOfProjectMax: 0,
-    isFilteringAgeOfProject: false,
-    price: [0, 100],
-    priceMax: 0,
-    isFilteringPrice: false,
-    isFilteringNFTOnly: false
-  }
-  const [filters, dispatch] = useReducer(filterReducer, initialState)
+  const {filters, dispatch } = useSearchFilter()
   const [toggleFilters, setToggleFilters] = useState(false)
   useEffect(() => {
     const handleResize = () => {
